Move JSONEncoder/Decoder from sqlite to helper
