+++ opsdroid/database/sqlite/__init__.py
@@ -2,11 +2,11 @@
 import os
 import logging
 import json
-import datetime
 import aiosqlite
 
 from opsdroid.const import DEFAULT_ROOT_PATH
 from opsdroid.database import Database
+from opsdroid.helper import JSONEncoder, JSONDecoder
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -118,122 +118,122 @@ class DatabaseSqlite(Database):
         return data
 
 
-class JSONEncoder(json.JSONEncoder):
-    """A extended JSONEncoder class.
-
-    This class is customised JSONEncoder class which helps to convert
-    dict to JSON. The datetime objects are converted to dict with fields
-    as keys.
-
-    """
-
-    # pylint: disable=method-hidden
-    # See https://github.com/PyCQA/pylint/issues/414 for reference
-
-    serializers = {}
-
-    def default(self, o):
-        """Convert the given datetime object to dict.
-
-        Args:
-            o (object): The datetime object to be marshalled.
-
-        Returns:
-            dict (object): A dict with datatime object data.
-
-        Example:
-            A dict which is returned after marshalling::
-
-                {
-                    "__class__": "datetime",
-                    "year": 2018,
-                    "month": 10,
-                    "day": 2,
-                    "hour": 0,
-                    "minute": 41,
-                    "second": 17,
-                    "microsecond": 74644
-                }
-
-        """
-        marshaller = self.serializers.get(type(o), super(JSONEncoder, self).default)
-        return marshaller(o)
-
-
-class JSONDecoder:
-    """A JSONDecoder class.
-
-    This class will convert dict containing datetime values
-    to datetime objects.
-
-    """
-
-    decoders = {}
-
-    def __call__(self, dct):
-        """Convert given dict to datetime objects.
-
-        Args:
-            dct (object): A dict containing datetime values and class type.
-
-        Returns:
-            object or dct: The datetime object for given dct, or dct if
-                             respective class decoder is not found.
-
-        Example:
-            A datetime object returned after decoding::
-
-                datetime.datetime(2018, 10, 2, 0, 41, 17, 74644)
-
-        """
-        if dct.get("__class__") in self.decoders:
-            return self.decoders[dct["__class__"]](dct)
-        return dct
-
-
-def register_json_type(type_cls, fields, decode_fn):
-    """Register JSON types.
-
-    This method will register the serializers and decoders for the
-    JSONEncoder and JSONDecoder classes respectively.
-
-    Args:
-        type_cls (object): A datetime object.
-        fields (list): List of fields used to store data in dict.
-        decode_fn (object): A lambda function object for decoding.
-
-    """
-    type_name = type_cls.__name__
-    JSONEncoder.serializers[type_cls] = lambda obj: dict(
-        __class__=type_name, **{field: getattr(obj, field) for field in fields}
-    )
-    JSONDecoder.decoders[type_name] = decode_fn
+# class JSONEncoder(json.JSONEncoder):
+#     """A extended JSONEncoder class.
+#
+#     This class is customised JSONEncoder class which helps to convert
+#     dict to JSON. The datetime objects are converted to dict with fields
+#     as keys.
+#
+#     """
+#
+#     # pylint: disable=method-hidden
+#     # See https://github.com/PyCQA/pylint/issues/414 for reference
+#
+#     serializers = {}
+#
+#     def default(self, o):
+#         """Convert the given datetime object to dict.
+#
+#         Args:
+#             o (object): The datetime object to be marshalled.
+#
+#         Returns:
+#             dict (object): A dict with datatime object data.
+#
+#         Example:
+#             A dict which is returned after marshalling::
+#
+#                 {
+#                     "__class__": "datetime",
+#                     "year": 2018,
+#                     "month": 10,
+#                     "day": 2,
+#                     "hour": 0,
+#                     "minute": 41,
+#                     "second": 17,
+#                     "microsecond": 74644
+#                 }
+#
+#         """
+#         marshaller = self.serializers.get(type(o), super(JSONEncoder, self).default)
+#         return marshaller(o)
+#
+#
+# class JSONDecoder:
+#     """A JSONDecoder class.
+#
+#     This class will convert dict containing datetime values
+#     to datetime objects.
+#
+#     """
+#
+#     decoders = {}
+#
+#     def __call__(self, dct):
+#         """Convert given dict to datetime objects.
+#
+#         Args:
+#             dct (object): A dict containing datetime values and class type.
+#
+#         Returns:
+#             object or dct: The datetime object for given dct, or dct if
+#                              respective class decoder is not found.
+#
+#         Example:
+#             A datetime object returned after decoding::
+#
+#                 datetime.datetime(2018, 10, 2, 0, 41, 17, 74644)
+#
+#         """
+#         if dct.get("__class__") in self.decoders:
+#             return self.decoders[dct["__class__"]](dct)
+#         return dct
 
 
-register_json_type(
-    datetime.datetime,
-    ["year", "month", "day", "hour", "minute", "second", "microsecond"],
-    lambda dct: datetime.datetime(
-        dct["year"],
-        dct["month"],
-        dct["day"],
-        dct["hour"],
-        dct["minute"],
-        dct["second"],
-        dct["microsecond"],
-    ),
-)
-
-register_json_type(
-    datetime.date,
-    ["year", "month", "day"],
-    lambda dct: datetime.date(dct["year"], dct["month"], dct["day"]),
-)
-
-register_json_type(
-    datetime.time,
-    ["hour", "minute", "second", "microsecond"],
-    lambda dct: datetime.time(
-        dct["hour"], dct["minute"], dct["second"], dct["microsecond"]
-    ),
-)
+# def register_json_type(type_cls, fields, decode_fn):
+#     """Register JSON types.
+#
+#     This method will register the serializers and decoders for the
+#     JSONEncoder and JSONDecoder classes respectively.
+#
+#     Args:
+#         type_cls (object): A datetime object.
+#         fields (list): List of fields used to store data in dict.
+#         decode_fn (object): A lambda function object for decoding.
+#
+#     """
+#     type_name = type_cls.__name__
+#     JSONEncoder.serializers[type_cls] = lambda obj: dict(
+#         __class__=type_name, **{field: getattr(obj, field) for field in fields}
+#     )
+#     JSONDecoder.decoders[type_name] = decode_fn
+#
+#
+# register_json_type(
+#     datetime.datetime,
+#     ["year", "month", "day", "hour", "minute", "second", "microsecond"],
+#     lambda dct: datetime.datetime(
+#         dct["year"],
+#         dct["month"],
+#         dct["day"],
+#         dct["hour"],
+#         dct["minute"],
+#         dct["second"],
+#         dct["microsecond"],
+#     ),
+# )
+#
+# register_json_type(
+#     datetime.date,
+#     ["year", "month", "day"],
+#     lambda dct: datetime.date(dct["year"], dct["month"], dct["day"]),
+# )
+#
+# register_json_type(
+#     datetime.time,
+#     ["hour", "minute", "second", "microsecond"],
+#     lambda dct: datetime.time(
+#         dct["hour"], dct["minute"], dct["second"], dct["microsecond"]
+#     ),
+# )
