+++ tests/test_helper.py
@@ -1,4 +1,6 @@
 import os
+import asyncio
+import datetime
 import tempfile
 import unittest
 import unittest.mock as mock
@@ -10,6 +12,8 @@ from opsdroid.helper import (
     convert_ipynb_to_script,
     extract_gist_id,
     get_opsdroid,
+    JSONEncoder,
+    JSONDecoder,
 )
 
 
@@ -67,3 +71,71 @@ class TestHelper(unittest.TestCase):
     def test_opsdroid(self):
         # Test that get_opsdroid returns None if no instances exist
         assert get_opsdroid() is None
+
+
+class TestJSONEncoder(unittest.TestCase):
+    """A JSON Encoder test class.
+
+    Test the custom json encoder class.
+
+    """
+
+    def setUp(self):
+        self.loop = asyncio.new_event_loop()
+
+    def test_datetime_to_dict(self):
+        """Test default of json encoder class.
+
+        This method will test the conversion of the datetime
+        object to dict.
+
+        """
+        type_cls = datetime.datetime
+        test_obj = datetime.datetime(2018, 10, 2, 0, 41, 17, 74644)
+        encoder = JSONEncoder()
+        obj = encoder.default(o=test_obj)
+        self.assertEqual(
+            {
+                "__class__": type_cls.__name__,
+                "year": 2018,
+                "month": 10,
+                "day": 2,
+                "hour": 0,
+                "minute": 41,
+                "second": 17,
+                "microsecond": 74644,
+            },
+            obj,
+        )
+
+
+class TestJSONDecoder(unittest.TestCase):
+    """A JSON Decoder test class.
+
+    Test the custom json decoder class.
+
+    """
+
+    def setUp(self):
+        self.loop = asyncio.new_event_loop()
+
+    def test_dict_to_datetime(self):
+        """Test call of json decoder class.
+
+        This method will test the conversion of the dict to
+        datetime object.
+
+        """
+        test_obj = {
+            "__class__": datetime.datetime.__name__,
+            "year": 2018,
+            "month": 10,
+            "day": 2,
+            "hour": 0,
+            "minute": 41,
+            "second": 17,
+            "microsecond": 74644,
+        }
+        decoder = JSONDecoder()
+        obj = decoder(test_obj)
+        self.assertEqual(datetime.datetime(2018, 10, 2, 0, 41, 17, 74644), obj)
